Update and clean up run_tests.sh
The run_tests.sh script fails when it is run multiple times on CentOS.
The `bindep` run returns an empty list of packages and then `yum`
exits with an error since no packages were provided to install.

This patch checks the length of the `bindep` output and skips the `yum`
installation when the package list is empty.

The patch also cleans up some of the old cruft left over from previous
scripts and avoids repetition.

Change-Id: Ia759c1623e9a0f7421c0506d824b144132042e39
Signed-off-by: Major Hayden <major@mhtx.net>